Moved X-MAP over X-KEY
Per RFC, if the X-MAP is found after the X-KEY, it is supposed to be encrypted, but rare are the clients that are actually following the implementation (other than Safari's native hls implementation). In order to prevent problems, moving it over the key and never encrypting the X-MAP seems like the idea solution.